scripts: change how scripts are defined
Rather than the script using the api to define when it runs, this is now
done in tlconfig.lua. The problem with the original solution was if you
wanted to run the script without running a command you had to jump
through a lot of hoops.

i.e. even updating the autogenerated docs for cyan itself becomes a
chore when you have to rebuild files when updating comments