breaking: Change how scripts are loaded
Scripts are now configured in `tlconfig.lua` by specifying the hook as
the key and an array of file names to execute in order. E.g. in our own
`tlconfig.lua`:
	scripts = {
		["build:post"] = {
		   "scripts/gen_rockspec.tl",
		   "scripts/docgen.tl",
		}
	}

This allows scripts to be ran multiple times and in a deterministic
order.

This comes with breaking changes to the `cyan.script` module listed
here:
 - instead of directly `load`ing scripts, the file path is registered
   for a given command and hook. Paths may be registered multiple times
   as this just appends to an array.
 - scripts are now actually _loaded_ via the `ensure_loaded_for_command`
   function
 - `emit_hook` will now assert that `ensure_loaded_for_command` was
   called before it

There is currently no way to unregister scripts