build: process all graph nodes at once, rather than staging by mark
Checking only nodes marked for type checking first could lead to
inconsistencies where a dependent file (which is only marked for type
checking) is checked before its parent (which is marked for compilation)
and the parent declares a global.

Since the dependent is pulled in first, it will implicitly pull in the
parent to be checked (declaring the global), then the parent will be
checked for its compilation (declaring the global a second time),
causing an incorrect error to be generated.

I'm not totally convinced that this completely solves the issue, but I
am not in the mood to do some graph theory and prove it. Given that the
graph is acyclic and nodes are processed in the order of their weight I
am happy to handwavingly believe that this should solve it.

Closes #36