Remove `module_name` feature
This feature was meant to solve the slight annoyance of having to make a
subdirectory with your module name for intra-project `require`s to find
the correct modules.

The implementation of this is a hack at best and as far as I know,
people aren't really using it. (Likely due to the discoverability of
config options to be lacking).