script: introduce a script cache for scripts written in teal
Writing a script in teal introduces a noticable delay since it needs to
be compiled each time. Using our patented
check-the-mod-time-of-the-files™®© technology we can compile it once and
shove it in some random directory to be pulled up later when it hasn't
changed.

This behavior can be controlled with a few environment variables:

 - CYAN_DISABLE_SCRIPT_CACHE: disables saving and loading to and from
   the cache

 - CYAN_SCRIPT_CACHE_DIR: controls what directory saving and loading are
   done to and from.

   The default values on unix-likes are:

      - $XDG_CACHE_HOME/cyan-script-cache
      - $HOME/.cache/cyan-script-cache

   And on windows:

      - %AppData%\Temp\cyan-script-cache

Errors while saving and loading to and from the cache are not considered
failures as it is trivial to just recompile the script.