Introduce `InvocationContext`
For script caching we started storing script paths as absolute, but
displayed paths are generally relative to the starting directory. This
was hacked in as `script.base_directory`.

Introducing the `InvocationContext`.

To avoid global state, things that need information about how cyan
itself was invoked should take this as a parameter. Right now it only
contains the initial directory and project root directory.